# docs/openapi.yaml
openapi: 3.1.0
info:
  title: gigastt REST API
  version: "1.0"
  description: |
    Offline and streaming speech-to-text REST API powered by GigaAM v3 (rnnt head by default).
    On-device Russian speech recognition via ONNX Runtime.

    ## Endpoints

    - **Health & discovery**: `GET /health`, `GET /ready`, `GET /v1/models`
    - **File transcription**: `POST /v1/transcribe` — upload audio, receive full transcript
    - **Streaming transcription**: `POST /v1/transcribe/stream` — upload audio, receive SSE stream of partial/final segments
    - **OpenAI-compatible**: `POST /v1/audio/transcriptions` — multipart `file` + `model` → `{"text":"..."}`
    - **Async jobs** (opt-in `--enable-jobs`): `POST /v1/jobs`, `GET/DELETE /v1/jobs/{id}`, result + SSE events
    - **Admin**: `POST /v1/admin/reload` — loopback-only model hot-reload
    - **WebSocket** (see `docs/asyncapi.yaml`): `GET /v1/ws`

    Prometheus metrics are optional (behind `--metrics`) and exposed on a
    separate listener (default `127.0.0.1:9090`, override via `--metrics-listen`),
    not on this API surface.

    ## Audio Formats

    Supported upload formats: WAV (PCM, IEEE float, G.711, G.722, GSM 06.10,
    MS/IMA ADPCM, RF64/RIFX/BW64/Wave64), MP3, M4A/AAC, OGG/Vorbis, OGG/Opus (`.opus`),
    WebM/Opus (what a browser's `MediaRecorder` produces), FLAC. Raw headerless
    telephony streams are accepted with
    `?codec=pcmu|pcma|g722&sample_rate=N` (see the transcribe parameters).
...

    ## Limits

...

...

...

...

...
    {
...

...
servers:
...
paths:
...
  /v1/transcribe:
    post:
      summary: Transcribe audio file
      description: |
        Upload an audio file and receive a complete transcription.
        Supports WAV (PCM, IEEE float, G.711, G.722, GSM 06.10, ADPCM, RF64), MP3, M4A/AAC,
        OGG/Vorbis, OGG/Opus (`.opus`), WebM/Opus, FLAC, and raw telephony
        streams via `?codec=` / `?sample_rate=`.

        The default response is JSON. Alternative export formats are available
        via the `format` query parameter: `txt`, `srt`, `vtt`, `md`. Use
        `download` to send `Content-Disposition: attachment`.

        Pass `segments=true` to add a cue-grouped `segments` array to the default
        JSON response (same boundaries as the `srt` / `vtt` exports); combined
        with `format=md` it switches Markdown to `### [mm:ss]` segment headers.
        `txt` / `srt` / `vtt` ignore `segments` (already flat / cue-based).

        Long files are decoded and transcribed in bounded overlapping windows,
        not as a single inference session, so peak memory stays roughly
        constant regardless of file length — there is no default duration
        limit. Start the server with `--max-audio-secs <N>` (env
        `GIGASTT_MAX_AUDIO_SECS`) to reject audio longer than `N` seconds with
        `413 Payload Too Large` and code `audio_too_long`; the default (`0`)
        is unlimited. The VAD file path and streamed OGG/Opus are windowed the
        same way and carry no ceiling of their own. Speaker diarization,
        `channels=split`, and raw telephony codecs hold the whole
        decoded buffer in memory and always enforce a fixed ~30-minute safety
        ceiling on those paths, independent of `--max-audio-secs`. G.722-in-WAV
        is windowed like other containers.
      operationId: transcribeFile
      parameters:
        - name: format
          in: query
          required: false
          schema:
            type: string
            enum: [json, txt, srt, vtt, md]
            default: json
          description: Export format for the transcription result.
        - name: download
          in: query
          required: false
          schema:
            type: string
          description: |
            Filename for the `Content-Disposition: attachment` header. If the
            value is empty, the server uses `transcript.<ext>`.
        - name: max_chars_per_line
          in: query
          required: false
          schema:
            type: integer
            default: 80
          description: Maximum characters per subtitle line (SRT/VTT). 0 = unlimited.
        - name: max_words_per_line
          in: query
          required: false
          schema:
            type: integer
            default: 14
          description: Maximum words per subtitle line (SRT/VTT). 0 = unlimited.
        - name: word_timestamps
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: Include per-word timestamps in Markdown output.
        - name: segments
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: |
            Add a `segments` array to the default JSON response. Segments are
            grouped from word timestamps by pause (≈0.9 s), sentence-ending
            punctuation (`.`, `!`, `?`), speaker change, or a maximum segment
            length of ≈30 s. With `format=md` the same segments are rendered as
            `### [mm:ss]` section headers. Ignored for `txt` / `srt` / `vtt`.
        - name: punctuation
          in: query
          required: false
          schema:
            type: boolean
          description: |
            Per-request override for the punctuation / casing restoration pass.
            `true` forces it on, `false` skips it; absent = the server's boot
            policy (byte-unchanged response). `true` when no punctuation model is
            loaded returns `409 punctuation_not_available`. Applies to this
            endpoint only — the SSE / WebSocket paths never run this pass.
        - name: itn
          in: query
          required: false
          schema:
            type: boolean
          description: |
            Per-request override for inverse text normalization (number-words →
            digits). `true` / `false` force the state; absent = boot policy. Pure
            code (no model), so always accepted. This endpoint only.
        - name: vad
          in: query
          required: false
          schema:
            type: boolean
          description: |
            Per-request override for VAD silence-skipping. `true` decodes only
            detected speech regions (requires the server started with `--vad`;
            otherwise `409 vad_not_loaded`), `false` decodes the whole buffer;
            absent = boot policy. This endpoint only.
        - name: variant
          in: query
          required: false
          schema:
            type: string
            enum: [rnnt, e2e_rnnt, ml_ctc, ml_ctc_large]
          description: |
            Forward-compatibility guard naming the recognition head the request
            expects. A single-model server can't switch heads, so any value
            other than the loaded variant returns `409 variant_not_loaded`;
            matching or absent proceeds. This endpoint only.
        - name: channels
          in: query
          required: false
          schema:
            type: string
            enum: [split]
          description: >-
            Use `split` to transcribe the left and right channels as separate
            speakers (`speaker_0` and `speaker_1`). Defaults to mono mix.
        - name: diarization
          in: query
          required: false
          schema:
            type: boolean
          description: >-
            Set to `true` to request polyvoice speaker diarization. The
            mutual-exclusion check with `channels=split` treats `diarization=true`
            as an explicit request; returns 400 with code `conflicting_modes` if
            both are set. Actual diarization output requires the speaker-encoder
            model to be loaded on the server (downloaded automatically when the
            `diarization` feature is enabled).
        - name: codec
          in: query
          required: false
          schema:
            type: string
            enum: [pcmu, pcma, g722, ulaw, alaw]
          description: >-
            Declare the upload body as a raw headerless telephony stream (RTP
            dump, Asterisk Monitor raw) instead of sniffing a container:
            `pcmu` (alias `ulaw`), `pcma` (alias `alaw`), or `g722`. Requires
            `sample_rate`; unknown values return 400 `unsupported_codec`.
            This endpoint only.
        - name: sample_rate
          in: query
          required: false
          schema:
            type: integer
          description: >-
            Sample rate (Hz) of a raw `codec` upload — mandatory when `codec`
            is set (typical telephony: 8000; missing or out-of-range values
            return 400 `invalid_sample_rate`). G.722 always decodes to its
            native 16 kHz; both 8000 (the SDP clock-rate convention) and 16000
            are accepted for it.
        - $ref: "#/components/parameters/Hotwords"
        - $ref: "#/components/parameters/HotwordsBoost"
      requestBody:
        required: true
        content:
          audio/wav:
            schema:
              type: string
              format: binary
          audio/mpeg:
            schema:
              type: string
              format: binary
          audio/mp4:
            schema:
              type: string
              format: binary
          audio/ogg:
            schema:
              type: string
              format: binary
          audio/flac:
            schema:
              type: string
              format: binary
          application/octet-stream:
            schema:
              type: string
              format: binary
            description: Raw audio bytes (format auto-detected)
      responses:
        "200":
          description: Transcription successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranscribeResponse"
              example:
                text: "привет как дела"
                words:
                  - word: "привет"
                    start: 0.5
                    end: 0.8
                    confidence: 0.97
                    speaker: 0
                  - word: "как"
                    start: 0.9
                    end: 1.0
                    confidence: 0.95
                    speaker: 1
                  - word: "дела"
                    start: 1.1
                    end: 1.4
                    confidence: 0.92
                    speaker: 0
                duration: 2.5
                confidence: 0.95
            text/plain:
              schema:
                type: string
              example: "привет как дела"
            application/x-subrip:
              schema:
                type: string
              example: |
                1
                00:00:00,000 --> 00:00:02,500
                привет как дела
            text/vtt:
              schema:
                type: string
              example: |
                WEBVTT

                00:00:00.000 --> 00:00:02.500
                привет как дела
            text/markdown:
              schema:
                type: string
              example: |
                ---
                duration: 2.5
                language: ru
                speakers: 0
                ---

                # Transcript

                привет как дела
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/OverrideConflict"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "422":
          $ref: "#/components/responses/TranscriptionError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /v1/transcribe/stream:
    post:
      summary: Transcribe audio file (SSE streaming)
      description: |
        Upload an audio file and receive a stream of transcription results
        via Server-Sent Events (SSE). Results are emitted as partial and final
        segments as the audio is processed chunk-by-chunk.

        Each SSE event contains a JSON object with `type` discriminator:
        - `partial` — interim result (may change)
        - `final` — finalized utterance
        - `error` — processing error
      operationId: transcribeStream
      requestBody:
        required: true
        content:
          audio/wav:
            schema:
              type: string
              format: binary
          audio/mpeg:
            schema:
              type: string
              format: binary
          audio/mp4:
            schema:
              type: string
              format: binary
          audio/ogg:
            schema:
              type: string
              format: binary
          audio/flac:
            schema:
              type: string
              format: binary
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        "200":
          description: SSE stream of transcription segments
          content:
            text/event-stream:
              schema:
                type: string
              example: |
                data: {"type": "partial", "text": "привет", "timestamp": 1712700000.123, "words": [{"word": "привет", "start": 0.5, "end": 0.8}]}

                data: {"type": "final", "text": "привет как дела", "timestamp": 1712700001.456, "words": [{"word": "привет", "start": 0.5, "end": 0.8}, {"word": "как", "start": 0.9, "end": 1.0}, {"word": "дела", "start": 1.1, "end": 1.4}]}
        "400":
          $ref: "#/components/responses/BadRequest"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "422":
          description: Audio could not be decoded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "Failed to decode audio file. Check format."
                code: invalid_audio
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  parameters:
...
    Hotwords:
      name: hotwords
      in: query
      schema:
        type: string
      description: >
        Per-request hotword phrases, comma-separated (`сбер,тинькофф`).
        Absent keeps the engine boot glossary; present with an empty value
        forces biasing off; non-empty replaces the glossary for this request
        only (capped at 64 phrases / 64 chars each — over the cap the request
        fails with 400 `too_many_hotwords` / `hotword_phrase_too_long`).
    HotwordsBoost:
      name: hotwords_boost
      in: query
      schema:
        type: number
        format: float
      description: >
        Additive logit boost for `hotwords`. Ignored unless `hotwords` is
        also present; defaults to 5.0.
  schemas:
...
    TranscribeResponse:
      type: object
      required:
        - text
        - words
        - duration
      properties:
        text:
          type: string
          example: "привет как дела"
          description: Full transcript text
        words:
          type: array
          items:
            $ref: "#/components/schemas/WordInfo"
          description: Word-level transcription with timestamps
        duration:
          type: number
          format: double
          example: 2.5
          description: Audio duration in seconds
        confidence:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          example: 0.95
          description: |
            Mean confidence across all words (duration-weighted average of
            `words[].confidence`): an average of per-word softmax scores, not a
            calibrated transcript probability. Omitted when no words were
            decoded.
        diarization:
          $ref: "#/components/schemas/DiarizationNotice"
        segments:
          type: array
          items:
            $ref: "#/components/schemas/Segment"
          description: |
            Transcript segments grouped from word timestamps. Present only when
            the request passed `segments=true`; omitted otherwise so existing
            clients are unaffected. Segments split on pause (≈0.9 s),
            sentence-ending punctuation (`.`, `!`, `?`), speaker change, or a
            maximum segment length of ≈30 s. With `format=md&segments=true`
            these same boundaries are rendered as `### [mm:ss]` sections.
...
    WordInfo:
      type: object
      required:
        - word
        - start
        - end
        - confidence
      properties:
        word:
          type: string
          example: "привет"
          description: Recognized word
        start:
          type: number
          format: double
          example: 0.5
          description: Word start time in seconds
        end:
          type: number
          format: double
          example: 0.8
          description: Word end time in seconds
        confidence:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          example: 0.97
          description: Softmax confidence score (0.0–1.0), averaged over the word's constituent tokens
        speaker:
          type: integer
          description: Zero-based speaker index (only present when channel split or diarization produced speaker labels)
...
    ErrorResponse:
      type: object
      required:
        - error
        - code
      properties:
        error:
          type: string
          example: "Request body exceeds the configured size limit"
          description: Human-readable error description
        code:
          type: string
          example: payload_too_large
          description: Machine-readable error code
        partial:
          $ref: "#/components/schemas/PartialTranscript"
        retry_after_ms:
          type: integer
          example: 30000
          description: |
            Suggested retry delay in milliseconds. Present only for transient
            backpressure errors (pool saturation).
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Empty request body"
            code: empty_body
    PayloadTooLarge:
      description: |
        Request body exceeds `--body-limit-bytes`, or decoded audio exceeds a
        configured duration limit (`--max-audio-secs`, or the fixed
        ~30-minute safety ceiling on whole-buffer paths: diarization,
        `channels=split`, telephony).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            payload_too_large:
              summary: Body exceeds --body-limit-bytes
              value:
                error: "Request body exceeds the configured size limit"
                code: payload_too_large
            audio_too_long:
              summary: Decoded audio exceeds a configured duration limit
              value:
                error: "Audio file too long. Maximum supported: 1800s."
                code: audio_too_long
    TranscriptionError:
      description: Transcription failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Transcription failed. Check audio format."
            code: transcription_error
    OverrideConflict:
      description: |
        A per-request override requested a knob the server can't honour: a knob
        turned on without its backing resource loaded, or a `variant` that isn't
        the loaded head. Returned before any pool checkout (fail-fast).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            vad_not_loaded:
              summary: "?vad=true but the server has no VAD"
              value:
                error: "VAD requested but not loaded; start the server with --vad"
                code: vad_not_loaded
            punctuation_not_available:
              summary: "?punctuation=true but no punctuation model is loaded"
              value:
                error: "punctuation requested but no punctuation model is loaded"
                code: punctuation_not_available
            variant_not_loaded:
              summary: "?variant=<head> differs from the loaded head"
              value:
                error: "Requested model variant is not loaded"
                code: variant_not_loaded
    ServiceUnavailable:
      description: Server busy or shutting down
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retry
          example: 30
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            timeout:
              summary: Pool saturated
              value:
                error: "Server busy, try again later"
                code: timeout
                retry_after_ms: 30000
            pool_closed:
              summary: Graceful shutdown in progress
              value:
                error: "Server is shutting down"
                code: pool_closed
    TooManyRequests:
      description: Per-IP rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retry
          example: 30
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Too many requests"
            code: rate_limited
            retry_after_ms: 30000
    GatewayTimeout:
      description: Inference timed out
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "Inference timed out."
            code: inference_timeout
...